// src/app/shared/models/channelpartner/view-profile.model.ts
export interface IchannelPartnerModel {
    AadharNo: number;
    AccountId: string;
    AccountNo: number;
    AddressLine1: string;
    AddressLine2: string;
    Agreement: string;
    AlternateNumber: number;
    Area: string;
    BankName: string;
    Branch: string;
    BrokerId: number;
    City: string;
    ContactId: string;
    Country: string;
    CrmAgent: number;
    DateOfBirth: string;
    Email: string;
    EntityId: number;
    GSTIN: number;
    GSTIN1: number;
    IFSC: number;
    Landmark: string;
    Level: string;
    MobileNo: number;
    Name: string;
    Occupation: string;
    Organization: string;
    PANType: number;
    PANVerificationStatus: string;
    PanNo: string;
    PanNo1: string;
    Pincode: number;
    ReferredBy: number;
    ReferredByName: string;
    ReportingTo: string;
    ReraEndDate: string;
    ReraNo: number;
    ReraStartDate: string;
    State: string;
    TAN: number;
    Type: string;
    UPICode: string;
    UserType: string;
    DocumentName: string;
    Gender: string;
    agreement: string;
    AgreementFlag: string;
}